Refactor RingLink, add TrapHelpers, energy fixes
refactor traps to reduce duplicate code, fixes for energy and ring links, uhhh im sure theres some badge logic fixed for badges but i forget which ones, will have to double check them all anyway

+++ AP_Links/EnergyLinkStore.cs
@@ -286,7 +286,7 @@ namespace Peak.AP
         private EnergyLinkService _energyLinkService;
         private BoxCollider _collider;
         private Animator _animator;
-        private int _cachedEnergy = 0;
+        private long _cachedEnergy = 0;
         private string _selectedBundleName;
         private Action _selectedBundleAction;
         private const int BUNDLE_COST = 400000000; // 400 MJ
